tzdev: check for insufficiently sized inputs.
* drivers/misc/sierra_tzdev.c (tzdev_generate_RSA_keypair): Function now
takes a size argument. When casting the rsa_params_ptr to the structure
type, it checks that the size is large enough that the indicated object
can contain that structure type. If the object isn't large enough, then
we access out of bounds (of kmalloced memory, not user memory).
(sierra_tzdev_ioctl): Pass the plain data size to
tzdev_generate_RSA_keypair.  Also, in the handling of the
TZDEV_IOCTL_RSA_IMPORTKEY_REQ, a similar issue is addressed. The input
buffer representing key material is converted to a structure, and then
the fields are referenced individually. That is only valid if the key
material is large enough to be that structure, otherwise we have
accesses out of bounds.

Change-Id: Ifde4334db71f7cbf8addc47550a72d832f7ad7f4
Resolves: LXSWIREF-958
Signed-off-by: Kaz Kylheku <kkylheku@sierrawireless.com>