swimcu: boot_source/enable doesn't take effect for PSM or ULPM fallback
This patch makes the node '/sys/module/swimcu/boot_source/enable' as
a SYSFS API for user to submit PSM/ULPM enable/disable requests. The
node attributes is changed from write-only to readable and writable
to pass the events to user-space PSM glue-logic to handle the request.
It also adds status updates to '/sys/module/swimcu/psm/status' for
ULPM-related operations, which is handled by the driver.

Resolves: QTI9X07-2639
Change-Id: I3c41cff02bf24d3e15357c81fcfc9da67edb6d07
Signed-off-by: Dylan Zhang <dyzhang@sierrawireless.com>