wl18xx: fake a non-removable device
In order to select the appropriate SDIO power management option,
TI recommends that WL18XX is configured as a non-removable device.
This option, however, assumes that the device is present on SDIO
bus at boot time, checks its status on boot and stops polling for
SDIO bus state changes.

This is inconvenient for applications that optionally use WiFi,
i.e. which power it up on-demand. These devices would not be
detected at boot and - due to SDIO bus not being polled - don't
have the chance to be detected later in the boot session.

Allow the TI WL18XX device to be removable, but once detected and
configured, treat it as non-removable until device is removed in
software. Force an SDIO scan after the device is removed to resume
polling.

Resolves: LXSWIREF-1056
Change-Id: I0d26b16dcd59943846036f716eefca89919c9007
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>