msm_thermal: increase threshold temperature which is to start stepping CPU down
In current design, the thershold tempeature is set as 60 degC. When the
tempeature near CPU reaches 60 degC during bootup, the CPU clock frequency
is stepped down to prevent the thermal breakdown. So, longer boot up time is
observed.

The solution increases the value to 70 degC so that the normal bootup time
can be maintained 10 more degC.

Resolves: QTI9X07-2679
Change-Id: I39ddd0a67fb906261cc136b77aa229f542e4f40e
Signed-off-by: Ken Lam <kwlam@sierrawireless.com>