gpio-swimcu: restore IRQ type when enabling IRQs
swimcu_irq_disable disables the IRQ for a given GPIO by overwriting the
previous configuration. When swimcu_irq_enable is subsequently used to
reenable the GPIO interrupt, the configuration that was set before the
call to swimcu_irq_disable should be reapplied. Instead, the disabled
setting is applied again since the previous configuration was overwritten.
This patch updates the gpio-swimcu driver to save the current IRQ settings
when disabling interrupts and reapply them when interrupts are enabled.

Resolves: QTI9X07-3462
Change-Id: I01dbfb9509f230032bdb1bdbb24523d024183ff4
Signed-off-by: Mitchell Nichols <MNichols@sierrawireless.com>