i2c-msm-v2: Recover I2C bus if blocked while waiting for xfer to finish
The i2c bus gets blocked when accessing an external device along with
Sierra MCU. When the devices are accessed through the /dev/i2c
interface the communication uses the i2c-msm-v2 Qualcomm driver while
communication with Sierra MCU uses i2c-core driver. During this
simultaneous access i2c bus is acquired by the slave and gets blocked.
While there is a recovery mechanism in i2c-msm-v2 driver this is not
called at appropriate times as the status of the registers in the
master gets updated by the time it is being checked for i2c bus error.
Checking for i2c bus errors while waiting for transfer to complete
with the cached status does not block the i2c bus.

Resolves: QTI9X07-3139
Change-Id: I948ad9cf3d071782d743394d0ddc4c264a93ed85
Signed-off-by: Libinesh Kumaran Mottammal Thiy <lkmthiy@sierrawireless.com>