SWIMCU: Add retry mechanism when mci_protocol_command fails
Whenever there is I2C ARB_LOST error, the MCU GPIO wakeup interrupt
stops working, and then all the data communication will
be lost and corrupted.  The I2C bus will be recovered by QTI9X07-3139.
However, the MCU doesn.t recovered so as the result the MCU no
response to the MCI commands. In the kernel code, swimcu-prot.c,
add an error recovery/retry mechanism so that whenever the MCI
command is not successful, it will resend the command again.

Resolves: QTI9X07-3621
Change-Id: I5784857d2fec70574d54e42e9518aaf2e56e2d17
Signed-off-by: Kenneth Cheng <kce@sierrawireless.com>