Device Tree: Modify the "vin-sel" value to 2 for PM GPIO 2
AT!PMGPIO?2 reuturns with ERROR, and AT!PMGPIO?n where n=1, 3-6 returns status
successfully

Found that the returned error is caused by the retreived voltage source with
value of 3, and it is outside the supported range (i.e. 0 - ATPMGPIO_VIN_MAX=2)
for GPIO=2. In contrast, the voltage source assigned for other GPIOs is within
the range. The PM GPIO is configured and initialised in Linux via a device
source tree file "mdm9607-wp76xx.dtsi", and it is found that the voltage source
"vin-sel" is wrongly set to 3

Modify the "vin-sel" value to 2 for PM GPIO 2 to allow a correct initialisation
by Linux

Resolves: QTI9X07-3429
Change-Id: I9bfb79bfae01dc1b13f5bb41a6729d48e4f3af14
Signed-off-by: Cliff Choi <cchoi@sierrawireless.com>