swimcu: fix race condition where device fails to wakeup from ULPM
Description:
If the swimcu driver receives multiple ULPM requests, a race condition
can occur where the wakeup timer is cleared and the MDM never wakes up.

When the driver receives a ULPM request, it sends the wakeup timeout
to the MCU. The system will then proceed to powerdown. If the driver
receives another ULPM request while the system is powering down, the
driver clears the MCU wakeup timer and the MDM never wakes up.

Solution:
If the system has begun the process of entering ULPM, ignore subsequent
requests to enter ULPM.

Resolves: QTI9X06-319
Change-Id: I172639204d7ca602c0d1d59610a5882ab554a610
Signed-off-by: Bilal Nurhusien <bnurhusien@sierrawireless.com>