usb: Suppress USB3503 Reset quirk
According to Qualcomm, USB3503 Reset quirk makes the device undergo
"reset from resume instead of normal resume from suspend, which might
be useful for some devices not working with resume." Qualcomm recommends
to remove this quirk if the device is working fine with normal resume.

With this quirk in place, the device sometimes is unable to resume after
a suspend due to the following HSIC errors.
usb 1-1: device not accepting address 2, error -110
usb 1-1: device descriptor read/64, error -110

Resolves: LXSWIREF-1798
Change-Id: Ie7c6c802de7c00be3632d309f025f745ede7586c
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>